move local storage logic to localStorage.js
